time loop & keyboard
When I tried to use the starting template, it took me an embarrasingly
long time to figure out how to do this. It's not complicated code once
you see it but I think not having to look up the APIs helps a bit as a
starting point.

I've not yet made a browser-based game, so there's a reasonable chance
this is a totally bad way to go about it. If so, please let me know!

I've tried to extract things the user might want to change to pure
functions "near" the proper game entrypoint (`component:game`), and
moved the side-effects to the bottom of the file, with the idea that in
most cases users of the template would barely ever need to go and look
at the code at the bottom.

Happy to change anything, and feel free to reject this PR if this is not
the direction you want to go into.